app/requirements.txt
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
argcomplete==3.6.2
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.1.0
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.11
botocore==1.40.11
Brlapi==0.8.6
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.2.1
cockpit @ file:///builddir/build/BUILD/cockpit-344-build/cockpit-344/tmp/wheel/cockpit-344-py3-none-any.whl#sha256=6c0da7c2cbfb1207f38f238b3011ff193c786540a0404088b52184bb35356ba0
ConfigArgParse==1.7.1
contourpy==1.3.3
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dnf==4.23.0
fastapi==0.116.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.59.1
fros==1.1
fsspec==2025.7.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.8
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.12.0
identify==2.6.12
idna==3.10
imutils==0.5.4
iniconfig==2.1.0
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
langdetect==1.0.9
langtable==0.0.69
libcomps==0.1.22
libdnf==0.74.0
libvirt-python==11.0.0
locust==2.38.1
locust-cloud==1.26.3
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
matplotlib==3.10.5
ml_dtypes==0.5.3
mplcursors==0.6
mpmath==1.3.0
msgpack==1.1.1
mypy==1.17.0
mypy_extensions==1.1.0
networkx==3.5
nftables==0.1
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.47
onnx==1.19.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
Paste==3.10.1
pathlib==1.0.1
pathspec==0.12.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
pre_commit==4.2.0
productmd==1.45
progressbar2==4.5.0
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pyclipper==1.3.0.post6
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.7
pydantic_core==2.33.2
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyparted==3.13.0
pypdfium2==4.30.0
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-augeas==1.2.0
python-dateutil==2.8.2
python-doctr==1.0.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-socketio==5.13.0
python-utils==3.9.0
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.15.10-build/kernel-6.15.10/linux-6.15.10-200.fc42.x86_64/tools/net/ynl
pyzmq==27.0.1
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
ruff==0.12.4
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==80.9.0
shapely==2.1.1
simple-websocket==1.1.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
systemd-python==235
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
virtualenv==20.32.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
xkbregistry==0.3
zope.event==5.1.1
zope.interface==7.2
